Add bus service client project & BusServiceClient struct & change IncomingRequest to BusRequest
