+++ ngtq_bus_service/src/receiver.rs
@@ -5,7 +5,7 @@ use ngtask_queue::{CategoryTask, IdTask, TaskQueue};
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
-enum IncomingRequest {
+enum BusRequest {
     PushTask(Task),
     PullTask(TaskIdentifier),
     Error(String)
@@ -97,15 +97,15 @@ fn start_receiving(socket_path: String, tx: Sender<UnixStream>) {
 }
 
 fn handle_incoming_request(incoming_request: String, wrapped_task_queue: &Arc<Mutex<TaskQueue>>) -> BusResponse {
-    let deserialized_request: IncomingRequest = match serde_json::from_str(&incoming_request) {
+    let deserialized_request: BusRequest = match serde_json::from_str(&incoming_request) {
         Ok(incoming_request) => incoming_request,
         Err(error) => return BusResponse { successful: false, error: error.to_string(), payload: String::new() }
     };
 
     match deserialized_request {
-        IncomingRequest::PushTask(task) => handle_push_request(task, wrapped_task_queue),
-        IncomingRequest::PullTask(task_identifier) => handle_pull_request(task_identifier, wrapped_task_queue),
-        IncomingRequest::Error(error) => BusResponse { successful: false, error: error, payload: String::new() }
+        BusRequest::PushTask(task) => handle_push_request(task, wrapped_task_queue),
+        BusRequest::PullTask(task_identifier) => handle_pull_request(task_identifier, wrapped_task_queue),
+        BusRequest::Error(error) => BusResponse { successful: false, error: error, payload: String::new() }
     }
 }
