+++ ngtq_bus_service_client/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "ngtq_bus_service_client"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
